Update BaseModel.xml
Release 2.7.0. Schema revised to ensure backwards compatibility. To do so, both of the NamingAuthority and ResourceType text elements were made optional (0) instead of required (1). Also, the list of earlier version numbers of SPASE are now added to the SPASE 2.7.0 XSD file to permit validating any XML descriptions regardless of the actual SPASE version number listed in the descriptions. For instance, xmllint, which is a unix command, can be used to validate a SPASE version 2.0.0 (or equivalently any SPASE version earlier than 2.7.0) description with the SPASE 2.7.0 XSD file. Note that online SPASE Validator still requires checking by using the version listed in a SPASE XML description. LFB 20240401.